pyproject.toml
[tool.poetry]
name = "safeagent"
version = "0.1.0"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
uvicorn = "*"
pydantic = "*"
pydantic-settings = "*"
pytest = "*"
requests = "^2.32.5"
pyjwt = "^2.10.1"
sqlalchemy = "^2.0.46"
psycopg2-binary = "^2.9.11"
openai = "^2.15.0"
